Combine Gabe's template and Andrew's comp/services
