+++ src/app/login/login.component.ts
@@ -1,30 +1,12 @@
 import { Component, OnInit, ViewEncapsulation, ChangeDetectionStrategy } from '@angular/core';
 import {trigger, state, animate, style, transition} from '@angular/core';
-import { LoginService } from './login.service';
 
 @Component({
   selector: 'app-login',
   templateUrl: './login.component.html',
-  styleUrls: ['./login.component.css'],
-  encapsulation: ViewEncapsulation.Emulated,
-  changeDetection: ChangeDetectionStrategy.Default
+  styleUrls: ['./login.component.css']
 })
 export class LoginComponent implements OnInit {
-  loginUsername: string;
-  loginPassword: string;
-
-  constructor(private loginService: LoginService) { }
-
-  ngOnInit() {
-  }
-
-  submitLogin = () => {
-    this.loginService.login(this.loginUsername, this.loginPassword)
-      .subscribe(res => {
-        this.loginUsername = '';
-        this.loginPassword = '';
-      }, err => {
-        console.log('err', err)
-      });
-  }
+  constructor() {}
+  ngOnInit() {}
 }
