+++ src/app/login/signup/signup.component.html
@@ -2,15 +2,15 @@
 
   <table style="width: 100%" cellspacing="3"><tr>
     <td><md-input-container style="width: 100%">
-      <input required md-input placeholder="Email">
+      <input required md-input [(ngModel)]='inputUsername' placeholder="Email">
     </md-input-container></td>
     <td><md-input-container style="width: 100%">
-      <input required md-input placeholder="Password">
+      <input required md-input [(ngModel)]='inputPassword' placeholder="Password">
     </md-input-container></td>
   </tr></table>
 
   <p>
-    <button md-raised-button color='warn' routerLink="/login/form" class='login' > SIGN UP </button>
+    <button md-raised-button color='warn' (click)='submitSignup()' routerLink="/login/form" class='login' > SIGN UP </button>
 
     <md-input-container width="100%">
       <textarea md-input placeholder="social media placeholder" value=""></textarea>
