+++ src/app/login/signup/signup.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { SignupService } from './signup.service';
 
 @Component({
     selector: 'app-signup-form',
@@ -6,4 +7,23 @@ import { Component, OnInit } from '@angular/core';
     styleUrls: ['./signup.component.css']
 })
 
-export class SignupComponent {}
+export class SignupComponent implements OnInit {
+  inputUsername: string;
+  inputPassword: string;
+
+  constructor(private signUpService: SignupService) { }
+
+  ngOnInit() {
+  }
+
+  submitSignup = () => {
+    this.signUpService.signUp(this.inputUsername, this.inputPassword)
+    .subscribe(res => {
+      this.inputUsername = '';
+      this.inputPassword = '';
+    }, err => {
+      console.log('err', err);
+    });
+  }
+
+}
