--- src/app/signup/signup.component.html
@@ -1,3 +1,0 @@
-<input placeholder="New username" [(ngModel)]="inputUsername">
-<input placeholder="New password" [(ngModel)]="inputPassword">
-<button (click)="submitSignup()">Sign up!</button>
